+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/handler/mapping/HandlerMapping.java
@@ -1,0 +1,7 @@
+package com.interface21.webmvc.servlet.mvc.handler.mapping;
+
+import jakarta.servlet.http.HttpServletRequest;
+
+public interface HandlerMapping {
+    Object getHandler(HttpServletRequest request);
+}
